拆分切片man页面，为未来做准备
