+++ shuoshuo-player/src/player/pages/index.js
@@ -1,11 +1,13 @@
 import HomePage from './home';
 import FavListPage from './fav_list';
 import DiscoveryPage from './discovery';
+import LiveSlicersPage from "@player/pages/live_slicers";
 
 export const Pages = {
     HomePage,
     FavListPage,
     DiscoveryPage,
+    LiveSlicersPage
 }
 
 export default Pages;
