+++ shuoshuo-player/src/styles/player.scss
@@ -407,41 +407,6 @@ body, html, #root {
         .discovery-search-box {
             margin: 0 auto;
         }
-        .discovery-slicer-human {
-            max-width: ($width-16 + $width-2 * 2) * 3;
-            margin: 0 auto;
-            .slicer-box-title {
-                text-align: center;
-                padding: $width-3 0;
-            }
-            .discovery-slicer-human-container {
-                display: flex;
-                flex-wrap: wrap;
-                .slicer-card {
-                    display: flex;
-                    border-radius: $width-1;
-                    width: $width-16;
-                    margin: $width-2;
-                    border: 1px solid $grey-8;
-                    box-sizing: border-box;
-                    padding: $width-2;
-                    overflow: hidden;
-
-                    .slicer-meta {
-                        margin-left: $width-2;
-                    }
-
-                    .slicer-name {
-                        overflow: hidden;
-                        word-break: break-all;
-                        white-space: nowrap;
-                        text-overflow: ellipsis;
-                        line-height: 28px;
-                        padding-left: $width-2;
-                    }
-                }
-            }
-        }
         &.mode-search {
             .discovery-search-bar {
                 padding: $width-10 0 !important;
@@ -453,6 +418,42 @@ body, html, #root {
             em.keyword {
                 color: $red;
                 font-style:normal
+            }
+        }
+    }
+
+    .live-slicer-human {
+        max-width: ($width-16 + $width-2 * 2) * 3;
+        margin: 0 auto;
+        .slicer-box-title {
+            text-align: center;
+            padding: $width-3 0;
+        }
+        .live-slicer-human-container {
+            display: flex;
+            flex-wrap: wrap;
+            .slicer-card {
+                display: flex;
+                border-radius: $width-1;
+                width: $width-16;
+                margin: $width-2;
+                border: 1px solid $grey-8;
+                box-sizing: border-box;
+                padding: $width-2;
+                overflow: hidden;
+
+                .slicer-meta {
+                    margin-left: $width-2;
+                }
+
+                .slicer-name {
+                    overflow: hidden;
+                    word-break: break-all;
+                    white-space: nowrap;
+                    text-overflow: ellipsis;
+                    line-height: 28px;
+                    padding-left: $width-2;
+                }
             }
         }
     }
